[#1611] Registration honeypot field
Add a field to the registration form that is hidden using CSS. This
way only bots will enter text into the field and we can thus detect
when it's not a human trying to sign up. When that is the case we
only redirect to register_complete.html without creating a user.